+++ src/lib/MarkdownViewer.svelte
@@ -850,7 +850,7 @@
 		ontoggleHome={toggleHome}
 		ononpenFileLocation={openFileLocation}
 		ontoggleLiveMode={toggleLiveMode}
-		ontoggleEdit={toggleEdit}
+		ontoggleEdit={() => toggleEdit()}
 		{isEditing}
 		ondetach={handleDetach}
 		ontabclick={() => (showHome = false)}
@@ -882,7 +882,7 @@
 		ontoggleHome={toggleHome}
 		ononpenFileLocation={openFileLocation}
 		ontoggleLiveMode={toggleLiveMode}
-		ontoggleEdit={toggleEdit}
+		ontoggleEdit={() => toggleEdit()}
 		{isEditing}
 		ondetach={handleDetach}
 		ontabclick={() => (showHome = false)}
@@ -898,7 +898,20 @@
 			<div class="markdown-container" style="zoom: {zoomLevel}%" onwheel={handleWheel} role="presentation">
 				{#if isEditing}
 					<div class="editor-wrapper">
-						<Editor bind:value={rawContent} language={editorLanguage} onsave={saveContent} />
+						<Editor
+							bind:value={rawContent}
+							language={editorLanguage}
+							onsave={saveContent}
+							onnew={handleNewFile}
+							onopen={selectFile}
+							onclose={closeFile}
+							onreveal={openFileLocation}
+							ontoggleEdit={() => toggleEdit()}
+							ontoggleLive={toggleLiveMode}
+							onhome={() => (showHome = true)}
+							onnextTab={() => tabManager.cycleTab('next')}
+							onprevTab={() => tabManager.cycleTab('prev')}
+							onundoClose={handleUndoCloseTab} />
 					</div>
 				{:else}
 					<article bind:this={markdownBody} contenteditable="false" class="markdown-body" bind:innerHTML={htmlContent} onscroll={handleScroll}></article>